# api/openapi.yaml
openapi: 3.0.3
info:
  title: Logistic Document API
  version: 1.0.0
  description: >
    REST API for managing customers, performers, vehicles, drivers, orders,
    audit log, and generated transport documents.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/performers:
    get:
      tags: [Performers]
      summary: List performers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Performer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PerformerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Performers]
      summary: Create performer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerformerRequest'
      responses:
        '200':
          description: Created performer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/performers/{id}:
    get:
      tags: [Performers]
      summary: Get performer by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Performer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Performers]
      summary: Update performer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerformerRequest'
      responses:
        '200':
          description: Updated performer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Performers]
      summary: Delete performer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/vehicles:
    get:
      tags: [Vehicles]
      summary: List vehicles
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Vehicle list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Vehicles]
      summary: Create vehicle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
      responses:
        '200':
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/vehicles/{id}:
    get:
      tags: [Vehicles]
      summary: Get vehicle by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Vehicles]
      summary: Update vehicle
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
      responses:
        '200':
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Vehicles]
      summary: Delete vehicle
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Orders]
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '200':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Orders]
      summary: Update order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Orders]
      summary: Delete order permanently
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Order deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/orders/{id}/complete:
    post:
      tags: [Orders]
      summary: Complete order (generate documents)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Order completed; documents generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/orders/{id}/documents:
    get:
      tags: [Documents]
      summary: List generated documents for order
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Generated document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeneratedDocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Not authenticated
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Access denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    ProblemDetail:
      type: object
      description: RFC 7807 Problem Details (application/problem+json)
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        errors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    PerformerRequest:
      type: object
      required: [shortName]
      properties:
        shortName:
          type: string
        fullName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        inn:
          type: string
          nullable: true
        bik:
          type: string
          nullable: true
        kpp:
          type: string
          nullable: true
        paymentAccount:
          type: string
          nullable: true
        corrAccount:
          type: string
          nullable: true
        requisites:
          type: string
          nullable: true
    PerformerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        shortName:
          type: string
        fullName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        inn:
          type: string
          nullable: true
        bik:
          type: string
          nullable: true
        kpp:
          type: string
          nullable: true
        paymentAccount:
          type: string
          nullable: true
        corrAccount:
          type: string
          nullable: true
        requisites:
          type: string
          nullable: true
...
    VehicleRequest:
      type: object
      required: [performerId]
      properties:
        performerId:
          type: integer
          format: int64
        brandModel:
          type: string
          nullable: true
        plateNumber:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        isDefault:
          type: boolean
          nullable: true
    VehicleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        performerId:
          type: integer
          format: int64
        brandModel:
          type: string
          nullable: true
        plateNumber:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        isDefault:
          type: boolean
    OrderCreateRequest:
      type: object
      required: [customerId, performerId]
      properties:
        customerId:
          type: integer
          format: int64
        performerId:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
          nullable: true
        driverId:
          type: integer
          format: int64
          nullable: true
    OrderUpdateRequest:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        performerId:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        driverId:
          type: integer
          format: int64
        orderDate:
          type: string
          format: date
        orderNumber:
          type: integer
          format: int32
        loadingPlace:
          type: string
        loadingContact:
          type: string
        unloadingPlace:
          type: string
        unloadingContact:
          type: string
        tripCount:
          type: integer
          format: int32
        pricePerTrip:
          type: number
          format: decimal
        totalPrice:
          type: number
          format: decimal
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        performerId:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
          nullable: true
        driverId:
          type: integer
          format: int64
          nullable: true
        orderNumber:
          type: integer
          format: int32
        orderDate:
          type: string
          format: date
        loadingPlace:
          type: string
        loadingContact:
          type: string
          nullable: true
        unloadingPlace:
          type: string
        unloadingContact:
          type: string
          nullable: true
        tripCount:
          type: integer
          format: int32
        pricePerTrip:
          type: number
          format: decimal
          nullable: true
        totalPrice:
          type: number
          format: decimal
          nullable: true
        templateVersion:
          type: integer
          format: int32
    GeneratedDocumentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documentType:
          type: string
          enum: [CONTRACT_APPLICATION, WAYBILL, ACT_OF_WORK]
        fileFormat:
          type: string
          enum: [PDF, DOCX]
        sha256:
          type: string
        createdAt:
          type: string
          format: date-time
...